Initial commit of multiecho interface
