+++ fmriprep/workflows/bold/base.py
@@ -23,6 +23,7 @@ from ...interfaces import (
     GiftiNameSource
 )
 
+from ...interfaces.multiecho import T2SMap
 from ...interfaces.reports import FunctionalSummary
 
 # Fieldmap workflows
@@ -36,7 +37,6 @@ from ..fieldmap import (
 from .confounds import init_bold_confs_wf
 from .hmc import init_bold_hmc_wf
 from .stc import init_bold_stc_wf
-from .t2s_map import init_bold_t2s_map_wf
 from .registration import init_bold_reg_wf
 from .resampling import (
     init_bold_surf_wf,
@@ -199,7 +199,6 @@ def init_func_preproc_wf(bold_file, ignore, freesurfer,
         * :py:func:`~fmriprep.workflows.bold.util.init_bold_reference_wf`
         * :py:func:`~fmriprep.workflows.bold.stc.init_bold_stc_wf`
         * :py:func:`~fmriprep.workflows.bold.hmc.init_bold_hmc_wf`
-        * :py:func:`~fmriprep.workflows.bold.t2s_map.init_bold_t2s_map_wf`
         * :py:func:`~fmriprep.workflows.bold.registration.init_bold_reg_wf`
         * :py:func:`~fmriprep.workflows.bold.confounds.init_bold_confounds_wf`
         * :py:func:`~fmriprep.workflows.bold.resampling.init_bold_mni_trans_wf`
@@ -332,12 +331,6 @@ def init_func_preproc_wf(bold_file, ignore, freesurfer,
                                    mem_gb=mem_gb['filesize'],
                                    omp_nthreads=omp_nthreads)
 
-    # T2* driven BOLD registration to T1w
-    bold_t2s_reg_wf = init_bold_t2s_reg_wf(name='bold_t2s_reg_wf',
-                                           tes=tes,
-                                           mem_gb=mem_gb['filesize'],
-                                           omp_nthreads=omp_nthreads)
-
     # mean BOLD registration to T1w
     bold_reg_wf = init_bold_reg_wf(name='bold_reg_wf',
                                    freesurfer=freesurfer,
@@ -356,9 +349,6 @@ def init_func_preproc_wf(bold_file, ignore, freesurfer,
         metadata=metadata,
         name='bold_confounds_wf')
     bold_confounds_wf.get_node('inputnode').inputs.t1_transform_flags = [False]
-### RESUME HERE
-    if multiecho:
-        pass
 
     workflow.connect([
         (inputnode, bold_reference_wf, [('bold_file', 'inputnode.bold_file')]),
